backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
psutil==5.9.6